Semmle: Fix possible infite loops caused by overflow (#3503)
Comparison between differently sized integers in loop conditions can cause
infinite loops. This can happen when doing something like this:

```c
int64 very_big = MAX_INT32 + 1;
for (int32 i = 0; i < very_big; i++) {
    // do something
}
// never reached because i overflows before it can reach the value of very_big
```

(cherry picked from commit 3f7c5a5cf628290d89503b72e6673e482e9d8f05)

+++ src/backend/distributed/worker/worker_partition_protocol.c
@@ -936,7 +936,7 @@ FilterAndPartitionTable(const char *filterQuery,
 
 	while (SPI_processed > 0)
 	{
-		for (int rowIndex = 0; rowIndex < SPI_processed; rowIndex++)
+		for (uint64 rowIndex = 0; rowIndex < SPI_processed; rowIndex++)
 		{
 			HeapTuple row = SPI_tuptable->vals[rowIndex];
 			TupleDesc rowDescriptor = SPI_tuptable->tupdesc;
